% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/self-calib.R
\name{approx_quantile_phi}
\alias{approx_quantile_phi}
\title{Given a vector of draws, return a vector of length S of phi which best approximates the CDF.}
\usage{
approx_quantile_phi(draws, S)
}
\arguments{
\item{draws}{a vector of sample draws from a distribution}

\item{S}{number of quantile points}
}
\value{
vector of phi which are quantile function values
}
\description{
Given a vector of draws, return a vector of length S of phi which best approximates the CDF.
}
