% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_attributes.R
\name{attribute_present}
\alias{attribute_present}
\title{Get attribute presence for a vector of variable names.}
\usage{
attribute_present(attribute, variable_names, var_attr)
}
\arguments{
\item{attribute}{the attribute to test}

\item{variable_names}{names of variable to test for the attribute}

\item{var_attr}{an object defining the attributes, see \link{variable-attributes}.}
}
\value{
a logical vector the same length as \code{variable_names}
}
\description{
Get attribute presence for a vector of variable names.
}
