% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_attributes.R
\name{attribute_present_stats}
\alias{attribute_present_stats}
\title{Check an attribute  in the \verb{$stats} element of the results
for presence of an attribute.}
\usage{
attribute_present_stats(attribute, x)
}
\description{
Check an attribute  in the \verb{$stats} element of the results
for presence of an attribute.
}
