% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary-calibration-tests.R
\name{brier_resampling_test}
\alias{brier_resampling_test}
\alias{brier_resampling_p}
\alias{binary_miscalibration}
\alias{miscalibration_resampling_p}
\alias{miscalibration_resampling_test}
\title{Binary calibration tests}
\usage{
brier_resampling_test(x, y, alpha = 0.05, B = 10000)

brier_resampling_p(x, y, B = 10000)

binary_miscalibration(x, y)

miscalibration_resampling_p(x, y, B = 10000)

miscalibration_resampling_test(x, y, alpha = 0.05, B = 10000)
}
\arguments{
\item{x}{the predicted success probabilities}

\item{y}{the actual observed outcomes (just 0 or 1)}

\item{alpha}{the type I error rate for the test}

\item{B}{number of boostrap samples for the null distribution}
}
\value{
\code{brier_resampling_test} and \code{miscalibration_resampling_test} return
an object of class \code{htest}, \code{brier_resampling_p} and \code{miscalibration_resampling_p}
return just the p-value (for easier use with automated workflows).
\code{binary_miscalibration} computes just the miscalibration component using
the PAV (pool adjacent violators) algorithm.
}
\description{
Dimitriadis et al. propose several tests based on
comparing actual predictions to predictions when the probabilities are
calibrated. This yields several possible tests of correctly calibrated
predictions (i.e. that the expected proportion of true values matches the
predicted probability).
}
\details{
The \code{brier_} functions represent a test based on brier score, while
the \code{miscalibration_} functions represent a test based on miscalibration.
In both cases we evaluate the null distribution via bootstrapping.
}
\references{
\if{html}{\out{<div class="sourceCode">}}\preformatted{T. Dimitriadis, T. Gneiting, & A.I. Jordan,
}\if{html}{\out{</div>}}

Stable reliability diagrams for probabilistic classifiers,
Proc. Natl. Acad. Sci. U.S.A. 118 (8) e2016191118,
https://doi.org/10.1073/pnas.2016191118 (2021).
}
