% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary-calibration.R
\name{binary_calibration_from_bp}
\alias{binary_calibration_from_bp}
\alias{plot_binary_calibration_diff}
\alias{plot_binary_calibration_diff.SBC_results}
\alias{plot_binary_calibration_diff.data.frame}
\alias{plot_binary_calibration}
\alias{plot_binary_calibration.SBC_results}
\alias{plot_binary_calibration.data.frame}
\title{Binary prediction calibration computation and visualisation.}
\usage{
binary_calibration_from_bp(
  bp,
  type = c("reliabilitydiag", "calibrationband"),
  alpha = 0.05,
  ...,
  region.position = NULL
)

plot_binary_calibration_diff(
  x,
  type = c("reliabilitydiag", "calibrationband"),
  alpha = 0.05,
  ...,
  region.position = NULL,
  prob_histogram = TRUE
)

\method{plot_binary_calibration_diff}{SBC_results}(res, ...)

\method{plot_binary_calibration_diff}{data.frame}(
  bp,
  type = c("reliabilitydiag", "calibrationband"),
  alpha = 0.05,
  ...,
  region.position = NULL,
  prob_histogram = TRUE
)

plot_binary_calibration(
  x,
  type = c("reliabilitydiag", "calibrationband"),
  alpha = 0.05,
  ...,
  region.position = NULL,
  prob_histogram = TRUE
)

\method{plot_binary_calibration}{SBC_results}(res, ...)

\method{plot_binary_calibration}{data.frame}(
  bp,
  type = c("reliabilitydiag", "calibrationband"),
  ...,
  region.position = NULL,
  prob_histogram = TRUE
)
}
\arguments{
\item{bp}{the binary probabilities --- typically obtained with
\code{\link[=binary_probabilities_from_stats]{binary_probabilities_from_stats()}}. Can however be manually constructed,
it needs to be a \code{data.frame} with columns \code{variable}, \code{prob} and \code{simulated_value}.}

\item{type}{the type of calibration uncertainty bands to compute, see details.}

\item{alpha}{the level associated with the confidence intervals reports}

\item{...}{additional arguments passed to
\code{\link[reliabilitydiag:reliabilitydiag]{reliabilitydiag::reliabilitydiag()}} or \code{\link[calibrationband:calibration_bands]{calibrationband::calibration_bands()}}}

\item{region.position}{for \code{type ="reliabilitydiag"} we may choose whether
the uncertainty interval surrounds the estimate (\code{region.position = "estimate"})
or the null distribution (\code{region.position = "diagonal"})}

\item{prob_histogram}{Whether a histogram of the observed probabilities should
be overlaid with the calibration curve.}

\item{res}{An \link{SBC_results} object}
}
\value{
\code{binary_calibration_from_bp} returns a \code{data.frame} with columns \code{variable}, \code{prob}, \code{estimate}, \code{low} and \code{high},
for each variable, it contains an estimate + confidence interval across a range
of probabilities (in equal steps). The \code{plot_} methods return a \code{ggplot2}
object showing either the calibration curve or the difference between
the calibration curve and perfect calibration (the diagonal)
}
\description{
Obtain estimate of binary prediction calibration and the associated
uncertainty interval for further processing or plot it directly.
}
\details{
When \code{type = "reliabilitydiag"}, the intervals are based on
\code{\link[reliabilitydiag:reliabilitydiag]{reliabilitydiag::reliabilitydiag()}} and depending on \code{region.position}
can be centered on the null distribution of perfect calibration or on the
estimated calibration.
When \code{type = "calibrationband"} the intervals
are around the estimated calibration using \code{\link[calibrationband:calibration_bands]{calibrationband::calibration_bands()}}
--- in our experience the \code{calibrationband}
method has less sensitivity to detect miscalibration, but they require
somewhat weaker assumptions.
}
