% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary-calibration.R
\name{binary_probabilities_from_stats}
\alias{binary_probabilities_from_stats}
\title{Extract binary probabilities from SBC stats.}
\usage{
binary_probabilities_from_stats(stats)
}
\value{
A \code{data.frame} with at least the columns \code{variable} and \code{prob}.
}
\description{
Takes all variables marked with \code{\link[=binary_var_attribute]{binary_var_attribute()}} in the
statistics (the \verb{$stats} field of an \code{SBC_results} object)
and transforms the reported CDF values into probabilities
that the underlying binary variable is 1.
}
\details{
To support exact binary probabilities, a backend must implement \code{\link[=SBC_posterior_cdf]{SBC_posterior_cdf()}},
as exact probabilities cannot be derived from samples.
}
