% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{bind_datasets}
\alias{bind_datasets}
\title{Combine multiple datasets together.}
\usage{
bind_datasets(...)
}
\arguments{
\item{...}{datasets to bind}
}
\description{
Combine multiple datasets together.
}
