% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derived-quantities.R
\name{bind_generated_quantities-deprecated}
\alias{bind_generated_quantities-deprecated}
\title{Combine two lists of derived quantities}
\description{
Delegates directly to \code{bind_derived_quantities()}.
}
\seealso{
\code{\link{SBC-deprecated}}
}
\keyword{internal}
