% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{bind_generated_quantities}
\alias{bind_generated_quantities}
\title{Combine multiple generated quantities into a single object}
\usage{
bind_generated_quantities(...)
}
\description{
Combine multiple generated quantities into a single object
}
