% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generator-brms.R
\name{brms_response_sequence}
\alias{brms_response_sequence}
\title{Determine the response sequence of brms model}
\usage{
brms_response_sequence(x)
}
\description{
Determine the response sequence of brms model
}
