% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backends.R
\name{brms_SBC_backend}
\alias{brms_SBC_backend}
\title{Build a backend based on the \code{brms} package.}
\usage{
brms_SBC_backend(..., template_dataset)
}
\arguments{
\item{...}{arguments passed to \code{brm}.}

\item{template_dataset}{a representative dataset that can be used to generate code.}
}
\description{
Build a backend based on the \code{brms} package.
}
