% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{brms_SBC_generator}
\alias{brms_SBC_generator}
\title{Create a brms generator.}
\usage{
brms_SBC_generator(..., generate_lp = TRUE)
}
\arguments{
\item{...}{arguments passed to \code{brms::brm}}

\item{generate_lp}{whether to compute the overall log-likelihood of the model
as an additional parameter. This can be somewhat computationally expensive,
but improves sensitivity of the SBC process.}
}
\description{
Brms generator uses a brms model with \code{sample_prior = "only"} to generate
new datasets.
}
