% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend-cached.R
\name{cached_fit_filename}
\alias{cached_fit_filename}
\title{Provide a file where a cached result of a fit will be/was stored.}
\usage{
cached_fit_filename(cache_dir, backend, generated)
}
\description{
This is to allow you to directly access the cache, should you need to.
The cache file will always be a \code{list} readable by \code{readRDS} and have
element \verb{$fit} for the actual fit. Other elements will contain the
text output, messages and warnings emitted by the fit.
}
