% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{calculate_prior_sd}
\alias{calculate_prior_sd}
\title{Calculate  prior standard deviation of a dataset}
\usage{
calculate_prior_sd(datasets)
}
\value{
a named vector of prior SDs
}
\description{
Calculate  prior standard deviation of a dataset
}
