% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{calculate_rank}
\alias{calculate_rank}
\title{Given prior and posterior samples, generate rank count for each bin.}
\usage{
calculate_rank(prior, posterior, thin)
}
\arguments{
\item{prior}{A named array of dimensions(n_iter, n_pars) where n_iter=number of SBC draws, n_pars the number of parameters. Names should be applied as parameter names.}

\item{posterior}{A named array of dimensions(n_iter, n_pars, n_samples) where n_iter=number of SBC draws, n_pars the number of parameters, and n_samples the number of samples per SBC draw. Names should be applied as parameter names. Equivalent to posterior.as_draws_array}

\item{thin}{Integer in which thinning samples will be applied
prior array dimensions: (n_iter, n_pars)
posterior array dimensions: (n_iter, n_pars, n_sample)}
}
\value{
array of dimensions(n_iter, n_pars)
}
\description{
Given prior and posterior samples, generate rank count for each bin.
}
