% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{calculate_rank_rvars}
\alias{calculate_rank_rvars}
\title{Given prior and posterior samples, generate rank count for each bin.}
\usage{
calculate_rank_rvars(prior, posterior, param)
}
\arguments{
\item{prior}{A named array of dimensions(n_iter, n_pars) where n_iter=number of SBC draws, n_pars the number of parameters. Names should be applied as parameter names.}

\item{posterior}{A named array of dimensions(n_samples, n_pars, n_iter) where n_iter=number of SBC draws, n_pars the number of parameters, and n_samples the number of samples per SBC draw. Names should be applied as parameter names. Equivalent to posterior.as_draws_array}

\item{param}{list of parameter names to calculate. If not given, calculate for all available parameters.}
}
\value{
array of dimensions(n_iter, n_pars)
}
\description{
Given prior and posterior samples, generate rank count for each bin.
}
