% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{calculate_ranks_draws_matrix}
\alias{calculate_ranks_draws_matrix}
\title{Calculate ranks given parameter values within a posterior distribution.}
\usage{
calculate_ranks_draws_matrix(params, dm)
}
\arguments{
\item{params}{a vector of values to check}

\item{dm}{draws_matrix of the fit (assumed to be already thinned if that was necessary)}
}
\description{
When there are ties (e.g. for discrete parameters), the rank is currently drawn stochastically
among the ties.
}
