% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{check_all_SBC_diagnostics}
\alias{check_all_SBC_diagnostics}
\title{Check diagnostics for an SBC results object.}
\usage{
check_all_SBC_diagnostics(results)
}
\value{
TRUE if all checks are OK, FALSE otherwise.
}
\description{
Produces a message for each failed check.
}
