% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{combine_all_variables}
\alias{combine_all_variables}
\alias{combine_array_elements}
\title{Helper functions to be passed to \link{ECDF-plots} to combine variables in a single
panel.}
\usage{
combine_all_variables(x)

combine_array_elements(x)
}
\arguments{
\item{x}{parameter names}
}
\description{
\code{combine_all_variables} will merge all variables in a single plot, while
\code{combine_array_elements} will merge all elements of any array into a single
panel of the plot
}
