% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{combine_args}
\alias{combine_args}
\title{Combine two named lists and overwrite elements with the same name
using the value from args2}
\usage{
combine_args(args1, args2)
}
\description{
Combine two named lists and overwrite elements with the same name
using the value from args2
}
