% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_attributes.R
\name{combine_var_attributes}
\alias{combine_var_attributes}
\title{Combine multiple sets of variable attributes.}
\usage{
combine_var_attributes(...)
}
\arguments{
\item{...}{the individual \code{\link[=var_attributes]{var_attributes()}} objects to combine.}
}
\description{
Combine multiple sets of variable attributes.
}
\details{
It is currently by design that multiple copies of an attribute are kept
}
