% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derived-quantities.R
\name{compute_dquants}
\alias{compute_dquants}
\title{Compute derived quantities based on given data and posterior draws.}
\usage{
compute_dquants(draws, generated, dquants, gen_quants = NULL)
}
\arguments{
\item{gen_quants}{Deprecated, use \code{dquants}}
}
\description{
Compute derived quantities based on given data and posterior draws.
}
