% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derived-quantities.R
\name{compute_gen_quants-deprecated}
\alias{compute_gen_quants-deprecated}
\title{Compute derived quantities based on given data and posterior draws.}
\description{
Delegates directly to \code{compute_dquants()}.
}
\seealso{
\code{\link{SBC-deprecated}}
}
\keyword{internal}
