% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{compute_results-deprecated}
\alias{compute_results-deprecated}
\title{Compute SBC results}
\description{
Delegates directly to \code{compute_SBC()}.
}
\seealso{
\code{\link{SBC-deprecated}}
}
\keyword{internal}
