% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{count_diagnostic}
\alias{count_diagnostic}
\title{A diagnostic representing count values.}
\usage{
count_diagnostic(
  label,
  error_above = 0,
  label_short = NULL,
  hint = "",
  error_only = FALSE
)
}
\arguments{
\item{label}{a description of the diagnostic}

\item{error_above}{a problem is signalled when the diagnostic is above this value}

\item{label_short}{optional shorter name for improved messages}

\item{hint}{a suggestion to the user to show when the diagnostic is problematic}

\item{error_only}{only report the diagnostic if at least one fit has a problematic value}
}
\description{
A diagnostic representing count values.
}
