% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{data_for_ecdf_plots}
\alias{data_for_ecdf_plots}
\title{Maybe not export in the end? Useful for debugging}
\usage{
data_for_ecdf_plots(x, ..., prob = 0.95, gamma = NULL, K = NULL)
}
\description{
Maybe not export in the end? Useful for debugging
}
