% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{default_chunk_size}
\alias{default_chunk_size}
\title{Determines the default chunk size.}
\usage{
default_chunk_size(n_fits, n_workers = future::nbrOfWorkers())
}
\description{
By default will make every worker process a single chunk.
You can set the \code{options(SBC.min_chunk_size = value)} to enforce a minimum
chunk size globally (chunk size can still be larger if you have substantially more
fits to run than workers.
}
