% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backends.R
\name{default_thin_ranks}
\alias{default_thin_ranks}
\title{S3 generic to get backend-specific default thinning for rank computation.}
\usage{
default_thin_ranks(backend)
}
\description{
Some backends (e.g. MCMC) require us to thin posterior samples before computing
ranks to remove autocorrelation. Other backends (e.g. ADVI/optimizing) already
produce independent samples, so need no thinning. Such backends can implement this
method to alter the default of \code{thin_ranks} argument for \code{\link[=compute_results]{compute_results()}}.
}
\details{
The default implementation plays it relatively safe and returns 10.
}
