% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backends.R
\name{diagnostic_types}
\alias{diagnostic_types}
\title{Get the types of diagnostics supported by a diagnostics data frame.}
\usage{
diagnostic_types(diags)
}
\arguments{
\item{diags}{a \code{data.frame} with the diagnostics (and an extra class to trigger
the correct S3 method)}
}
\value{
A named list where names are columns in the data frame and values
are objects that support the \code{get_diagnostic_messages_single()} function.
}
\description{
A backend may optionally implement this method to provide information
how the diagnostics reported by the backend can be best summarised and
presented to the user. See \code{vignette("implementing_backends")} for an
example implementation.
}
