% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derived-quantities.R
\name{dquants_var_attributes}
\alias{dquants_var_attributes}
\title{Get the \code{\link[=var_attributes]{var_attributes()}} object associated with the derived quantities}
\usage{
dquants_var_attributes(dquants)
}
\description{
Get the \code{\link[=var_attributes]{var_attributes()}} object associated with the derived quantities
}
