% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_ecdf}
\alias{plot_ecdf}
\alias{plot_ecdf_diff}
\title{Plot the ECDF-based plots.}
\usage{
plot_ecdf(
  x,
  variables = NULL,
  K = NULL,
  gamma = NULL,
  prob = 0.95,
  size = 1,
  alpha = 0.33,
  combine_variables = NULL,
  ecdf_alpha = NULL,
  ...,
  parameters = NULL
)

plot_ecdf_diff(
  x,
  variables = NULL,
  K = NULL,
  gamma = NULL,
  prob = 0.95,
  size = 1,
  alpha = 0.33,
  combine_variables = NULL,
  ecdf_alpha = NULL,
  ...,
  parameters = NULL
)
}
\arguments{
\item{x}{object supporting the \code{\link[=data_for_ecdf_plots]{data_for_ecdf_plots()}} method.}

\item{variables}{optional subset of variables to show in the plot}

\item{K}{number of uniformly spaced evaluation points for the ECDF or ECDFs. Affects
the granularity of the plot and can significantly speed up the computation
of the simultaneous confidence bands. Default value is chosen heuristically.
You can also use \code{"max"} to represent
the number of ranks or \code{"min"} to choose a lower but still sensible value.}

\item{gamma}{TODO}

\item{prob}{the width of the plotted confidence interval for the ECDF.}

\item{size}{size passed to \code{\link[ggplot2:geom_ribbon]{ggplot2::geom_ribbon()}} for the confidence band}

\item{alpha}{alpha level of the confidence band}

\item{combine_variables}{optionally specify a named list where each entry is a character
vectors which specifies a group of variables that will be displayed in the
same panel. Panel title will be the name of the list element.
A function that takes a character vector as an input and produces a list
can also be specified (see \link{combine-functions}).}

\item{ecdf_alpha}{the alpha level of the empirical CDF. Can be either a single number or
a function taking the number of variables that were combined (when \code{combine_variables}
is specified) and returns a number. By default, plots showing many
ECDFs will have reduced alpha.}

\item{...}{additional arguments passed to \code{\link[=data_for_ecdf_plots]{data_for_ecdf_plots()}}.
Most notably, if \code{x} is matrix, a \code{max_rank} parameter needs to be given.}

\item{parameters}{DEPRECATED, use \code{variables} instead.}
}
\description{
See \code{vignette("rank_visualizations")} for
more details.
See the methods for \code{\link[=data_for_ecdf_plots]{data_for_ecdf_plots()}} for available data formats.
}
\details{
\href{https://arxiv.org/abs/1903.08008}{arxiv::1903.08008} by A. Vehtari et al.
}
\seealso{
\code{\link[=plot_coverage]{plot_coverage()}}
}
