% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_attributes.R
\name{extract_attribute_arguments_stats}
\alias{extract_attribute_arguments_stats}
\title{Extract arguments of a given attribute from the \verb{$stats} element of the results}
\usage{
extract_attribute_arguments_stats(attribute_name, x)
}
\description{
Extract arguments of a given attribute from the \verb{$stats} element of the results
}
