% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_ppred.R
\name{full_ppred}
\alias{full_ppred}
\title{Full forward sampling of the PPC}
\usage{
full_ppred(fit, newdata = NULL, draws = NULL)
}
\arguments{
\item{fit}{An object of class \code{brmsfit}}

\item{newdata}{An optional data.frame for which to evaluate predictions. If NULL (default), the original data of the model is used.}

\item{draws}{An integer vector specifying the posterior draws to be used. If NULL (the default), all draws are used.}
}
\value{
A list of data.frames containing the draws.
}
\description{
Full forward sampling of the PPC
}
\examples{
# Pending
}
