% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{function_SBC_generator}
\alias{function_SBC_generator}
\title{Generate datasets via a function that creates a single dataset.}
\usage{
function_SBC_generator(f, ...)
}
\arguments{
\item{f}{function returning a list with elements \code{parameters}
(prior draws, a list or anything that can be converted to \code{draws_rvars}) and
\code{generated} (observed dataset, ready to be passed to backend)}

\item{...}{Additional arguments passed to \code{f}}
}
\description{
Generate datasets via a function that creates a single dataset.
}
