% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary-calibration-tests.R
\name{gaffke_test}
\alias{gaffke_test}
\alias{gaffke_test_paired}
\alias{gaffke_ci}
\alias{gaffke_p}
\title{Non-parametric test for the mean of a bounded variable.}
\usage{
gaffke_test(
  x,
  mu = 0.5,
  alpha = 0.05,
  lb = 0,
  ub = 1,
  B = 10000,
  alternative = c("two.sided", "less", "greater")
)

gaffke_test_paired(
  x,
  y,
  mu = 0,
  alpha = 0.05,
  lb = 0,
  ub = 1,
  B = 10000,
  alternative = c("two.sided", "less", "greater")
)

gaffke_ci(probs, B = 10000, alpha = 0.05)

gaffke_p(
  probs,
  mu = 0.5,
  alpha = 0.05,
  B = 10000,
  alternative = c("two.sided", "less", "greater")
)
}
\arguments{
\item{x}{a vector of observed values}

\item{mu}{the mean under null hypothesis}

\item{alpha}{the level of the test}

\item{lb, ub}{the lower and upper bounds for \code{x}
(for the paired version, it also applies to \code{y} you can provide a v
ector of length 2 to give different bounds to \code{x} and \code{y})}

\item{ub}{the upper bound for \code{x}}

\item{B}{number of bootstrap samples for the null distribution}

\item{alternative}{the alternative for the test.}

\item{y}{a second vector, the same length as x}

\item{probs}{a vector of observed values, scaled to [0, 1]}
}
\value{
\code{gaffke_test} and \code{gaffke_test_paired} return an object of class \code{htest}, \code{gaffke_p} and
\code{gaffke_ci} return just the p-value / CI as numeric for easier use in batch
workflows.
}
\description{
A non-parametric one-sample test for the mean.
The only assumptions it makes about the data is that they are i.i.d. and
that the underlying distribution is bounded from both sides.
}
\details{
The test is based on Gaffke 2005, though a more detailed analysis and
exposition can be found in Learned-Miller & Thomas 2020.

The test is expected to be valid for any bounded distribution without further
assumptions. The test has been proven valid only for special cases but
no counterexample is known despite some efforts in the literature to find
some. The test also provides way more power (and tighter confidence intervals)
than other non-parametric tests for bounded means - in fact its power converges
quite quickly to that of the t-test, which is known to be optimal
in the large data limit.

The paired two-sample version translates to a one-sample test with bounds
computed to constrain the difference. There is no known unpaired two-sample
version.

In SBC the test is useful for testing the data-averaged posterior
for binary variables - the paired version is to be used when the prior
is not analytically known and so the simulated value of the variable is the
second sample.

\code{gaffke_p} and \code{gaffke_ci} are intended for easier use in batch
workflows as they directly  return the relevant numeric values,
not a test object. Note that those functions expected their arguments
to be scaled to [0, 1] already, while \code{gaffke_test} will work
with any upper and lower bounds and do the scaling for you.
}
\examples{
set.seed(84623)
# true mean of x = 0.5
x <- rbeta(10, shape1 = 0.1, shape2 = 0.1)
gaffke_test(x)
# Standard t-test is invalid here due to small sample size
t.test(x, mu = 0.5)

# Paired version + scaling - true difference in means is ~2.5
x_scaled <- x * 2 + 1
y_scaled <- -1 + 3 * rbeta(length(x), shape1 = 5 * (x / 3), shape2 = 5 * (1 - x / 3))
gaffke_test_paired(x_scaled, y_scaled, lb = c(1, -1), ub = c(3, 2), mu = 2)

# As the number of draws increases, the test converges to a t-test
x <- rbeta(100, shape1 = 0.1, shape2 = 0.1)
gaffke_test(x)
t.test(x, mu = 0.5)

}
\references{
Gaffke, N. (2005).
“Three test statistics for a nonparametric one-sided hypothesis on the mean
of a nonnegative variable.” Mathematical Methods of Statistics, 14(4): 451–467.

Learned-Miller, E. and Thomas, P. S. (2020).
“A New Confidence Interval for the Mean of a Bounded Random Variable.”
https://arxiv.org/abs/1905.06208
}
