% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{generate_datasets}
\alias{generate_datasets}
\title{Generate datasets.}
\usage{
generate_datasets(generator, n_datasets)
}
\value{
object of class SBC_datasets
TODO: seed
}
\description{
Generate datasets.
}
