% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derived-quantities.R
\name{generated_quantities-deprecated}
\alias{generated_quantities-deprecated}
\title{Create a definition of derived quantities evaluated in R.}
\description{
Delegates directly to \code{derived_quantities()}.

Delegates directly to \code{validate_derived_quantities()}.
}
\seealso{
\code{\link{SBC-deprecated}}

\code{\link{SBC-deprecated}}
}
\keyword{internal}
