% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{get_diagnostic_messages}
\alias{get_diagnostic_messages}
\title{Get diagnostic messages for \code{SBC_results} or other objects.}
\usage{
get_diagnostic_messages(x)
}
\value{
An object of class \code{SBC_diagnostic_messages}, inheriting a data.frame.
}
\description{
Get diagnostic messages for \code{SBC_results} or other objects.
}
