% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{get_diagnostic_messages_single}
\alias{get_diagnostic_messages_single}
\title{Get diagnostic messages for a single diagnostic.}
\usage{
get_diagnostic_messages_single(diagnostic, values)
}
\arguments{
\item{diagnostic}{object describing the diagnostic}

\item{values}{the values of the diagnostics}
}
\value{
a \code{data.frame} with fields \code{type} (possible values \code{bad}, \code{info}, \code{ok}) and \code{message}
}
\description{
Get diagnostic messages for a single diagnostic.
}
\seealso{
\code{\link[=numeric_diagnostic]{numeric_diagnostic()}}, \code{\link[=count_diagnostic]{count_diagnostic()}}, \code{\link[=logical_diagnostic]{logical_diagnostic()}}
}
