% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_factors.R
\name{get_stats_for_submodel}
\alias{get_stats_for_submodel}
\title{Extract stats for a submodel from a Bayes factor results}
\usage{
get_stats_for_submodel(stats, submodel_id)
}
\description{
Extract stats for a submodel from a Bayes factor results
}
