% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{guess_bins}
\alias{guess_bins}
\title{Guess the number of bins for \code{\link[=plot_rank_hist]{plot_rank_hist()}}.}
\usage{
guess_bins(max_rank, N)
}
\arguments{
\item{max_rank}{the maximum rank observed}

\item{N}{the number of ranks observed}
}
\description{
Guess the number of bins for \code{\link[=plot_rank_hist]{plot_rank_hist()}}.
}
