% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{invtf_param}
\alias{invtf_param}
\title{Inverse transform parameters from uncontrained to constrained}
\usage{
invtf_param(param, tf)
}
\arguments{
\item{param}{uncontrained parameter \code{draws_rvars}}

\item{tf}{list of transformation type}
}
\value{
contrained parameter \code{draws_rvars}
}
\description{
Inverse transform parameters from uncontrained to constrained
}
