% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{invtf_param_vec}
\alias{invtf_param_vec}
\title{Inverse transform parameters from uncontrained to constrained}
\usage{
invtf_param_vec(param, tf)
}
\arguments{
\item{param}{a vector}

\item{tf}{string indicating transformation type}
}
\value{
constrained parameter vector
}
\description{
Inverse transform parameters from uncontrained to constrained
}
