% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/self-calib.R
\name{iter_stop}
\alias{iter_stop}
\title{Judge whether the SBC iteration have converged}
\usage{
iter_stop(param, param_next, target_vars, scm_init)
}
\arguments{
\item{param}{numeric vector of parameter values to be tested a.k.a prior sample}

\item{param_next}{numeric vector of updated parameter values after one SBC cycle a.k.a posterior values}

\item{target_vars}{function of parameters with which SBC iteration convergence are judged, null is strict testing all parameters}

\item{scm_init}{initial distance between initial parameter and its proposal}
}
\value{
boolean, judge convergence based on scm-scm_init ratio (relative) or \code{param}-\code{param_next} distance (absolute)
}
\description{
Judge whether the SBC iteration have converged
}
