% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{logical_diagnostic}
\alias{logical_diagnostic}
\title{A diagnostic representing binary values.}
\usage{
logical_diagnostic(
  ok_value,
  true_label,
  false_label = paste0("not ", true_label),
  hint = "",
  error_only = FALSE
)
}
\arguments{
\item{ok_value}{which value (\code{TRUE} or \code{FALSE}) signals a problem with the fit}

\item{true_label}{label to use for positive cases}

\item{false_label}{label to use for negative cases}

\item{hint}{a suggestion to the user to show when the diagnostic is problematic}

\item{error_only}{only report the diagnostic if at least one fit has a problematic value}
}
\description{
A diagnostic representing binary values.
}
