% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric.R
\name{max_diff}
\alias{max_diff}
\title{Max difference between binned samples with the same length}
\usage{
max_diff(x, y)
}
\arguments{
\item{x}{numeric vector of density from first distribution}

\item{y}{numeric vector of density from second distribution}

\item{...}{unused}
}
\value{
distance value based on max difference
}
\description{
Max difference between binned samples with the same length
}
