% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{numeric_diagnostic}
\alias{numeric_diagnostic}
\title{A diagnostic representing numerical values.}
\usage{
numeric_diagnostic(
  label,
  report = NULL,
  error_above = Inf,
  error_below = -Inf,
  allow_na = FALSE,
  label_short = NULL,
  hint = "",
  digits = 2,
  unit = ""
)
}
\arguments{
\item{label}{a description of the diagnostic}

\item{report}{what value shuld be reported to represent the diagnostic.
Support options are \code{"min"}, \code{"max"} and \code{"quantiles"}}

\item{error_above}{signal a problem when the diagnostic is above this value}

\item{error_below}{signal a problem when the diagnostic is below this value}

\item{allow_na}{if \code{FALSE}, any NA values are reported as a problem}

\item{label_short}{a shorter version of the label (if the original is long) for
cleaner summaries.}

\item{hint}{A suggestion what the user can do if problems are signalled}

\item{digits}{decimal digits to report}

\item{unit}{optional unit to report alongside the values}
}
\description{
Optionally the message can signal failure if the diagnostic is above/below
a certain threshold or is NA.
}
