% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.ecdf}
\alias{plot.ecdf}
\title{Plot ECDF given rank array "ranks"
\href{https://arxiv.org/abs/1903.08008}{arxiv::1903.08008} by A. Vehtari et al.}
\usage{
\method{plot}{ecdf}(ranks, par)
}
\arguments{
\item{ranks}{array of dimension(n_iter, n_pars) where n_iter=number of posterior draw iterations, n_pars the number of parameters of interest}

\item{par}{names of parameter to plot}
}
\description{
Plot ECDF given rank array "ranks"
\href{https://arxiv.org/abs/1903.08008}{arxiv::1903.08008} by A. Vehtari et al.
}
