% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.hist}
\alias{plot.hist}
\title{Plot Rank Histogram given rank array "ranks"}
\usage{
\method{plot}{hist}(ranks, par, thin, bins = 20)
}
\arguments{
\item{ranks}{array of dimension(n_iter, n_pars) where n_iter=number of posterior draw iterations, n_pars the number of parameters of interest}

\item{par}{names of parameter to plot}

\item{thin}{integer in which thinning was applied}

\item{bins}{number of histogram bins to plot default is 20}
}
\description{
Plot Rank Histogram given rank array "ranks"
}
