% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_contraction}
\alias{plot_contraction}
\title{Prior/posterior contraction plot}
\usage{
plot_contraction(x, prior_sd, parameters = NULL, scale = "sd", alpha = 0.8)
}
\description{
Prior/posterior contraction plot
}
