% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_coverage}
\alias{plot_coverage}
\alias{plot_coverage.SBC_results}
\alias{plot_coverage.data.frame}
\alias{plot_coverage_diff}
\alias{plot_coverage_diff.SBC_results}
\alias{plot_coverage_diff.data.frame}
\title{Plot the observed coverage and its uncertainty}
\usage{
plot_coverage(
  x,
  variables = NULL,
  prob = 0.95,
  interval_type = "central",
  parameters = NULL
)

\method{plot_coverage}{SBC_results}(
  x,
  variables = NULL,
  prob = 0.95,
  interval_type = "central",
  parameters = NULL
)

\method{plot_coverage}{data.frame}(
  x,
  variables = NULL,
  prob = 0.95,
  interval_type = "central",
  parameters = NULL
)

plot_coverage_diff(
  x,
  variables = NULL,
  prob = 0.95,
  interval_type = "central",
  parameters = NULL
)

\method{plot_coverage_diff}{SBC_results}(x, variables = NULL, prob = 0.95, interval_type = "central")

\method{plot_coverage_diff}{data.frame}(
  x,
  variables = NULL,
  prob = 0.95,
  interval_type = "central",
  parameters = NULL
)
}
\arguments{
\item{x}{object containing results (a data.frame or \code{\link[=SBC_results]{SBC_results()}} object).}

\item{variables}{variables to show in the plot or \code{NULL} to show all}

\item{prob}{the with of the uncertainty interval to be shown}

\item{parameters}{DEPRECATED. Use \code{variables} instead.}
}
\value{
a ggplot2 plot object
}
\description{
Please refer to \code{\link[=empirical_coverage]{empirical_coverage()}} for details on computation
and limitations of this plot as well as details on the arguments.
See \code{vignette("rank_visualizations")} for
more details.
}
