% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_sim_estimated}
\alias{plot_sim_estimated}
\title{Plot the simulated "true" values versus posterior estimates}
\usage{
plot_sim_estimated(x, parameters = NULL, estimate = "mean", alpha = 0.8)
}
\description{
Plot the simulated "true" values versus posterior estimates
}
