% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/self-calib.R
\name{quantile_huber_loss}
\alias{quantile_huber_loss}
\title{Calculate qualtile huber loss for a given tau tau_{s_index}}
\usage{
quantile_huber_loss(phi_prior, phi_post, s_index, k, S, n_post_samples)
}
\arguments{
\item{phi_prior}{vector of phis for the prior(pre-transformation) distribution}

\item{phi_post}{vector of phis for the posterior(post-transformation) distribution}

\item{s_index}{The tau index to calculate loss}

\item{k}{interval to calculate loss. resulting loss will be in range [-k, k]}

\item{S}{the number of phis. equal to length(phi_prior) = length(phi_post)}

\item{n_post_samples}{the number of samples to draw from posterior, to approximate the expected huber loss}
}
\value{
a vector of length 2, where the first value is the expected huber loss and the second value the number of posterior samples less than the quantile value at phi[s_index]
}
\description{
Calculate qualtile huber loss for a given tau tau_{s_index}
}
