% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric.R
\name{rank2unif}
\alias{rank2unif}
\title{Distance between binned draws (rank for SBC) and discrete uniform}
\usage{
rank2unif(results, par, bins = 20)
}
\arguments{
\item{par}{names of parameter to plot}

\item{bins}{number of bins to use for summary}

\item{ranks}{array of dimension (n_iter, n_pars) where n_iter=number of posterior draw iterations, n_pars the number of parameters of interest}

\item{thin}{integer in which thinning was applied}
}
\description{
Distance between binned draws (rank for SBC) and discrete uniform
}
