% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{rank_summary}
\alias{rank_summary}
\title{Summarize Ranks on different distance maeasure metrics}
\usage{
rank_summary(ranks, par, thin, bins = 20)
}
\arguments{
\item{ranks}{array of dimension(n_iter, n_pars) where n_iter=number of posterior draw iterations, n_pars the number of parameters of interest}

\item{par}{names of parameter to plot}

\item{thin}{integer in which thinning was applied}

\item{bins}{number of bins to use for summary}
}
\description{
Summarize Ranks on different distance maeasure metrics
}
