% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend-bridgesampling.R
\name{rbind.SBC_bridgesampling_diagnostics}
\alias{rbind.SBC_bridgesampling_diagnostics}
\title{Custom rbind implementation maintainig information about submodels}
\usage{
\method{rbind}{SBC_bridgesampling_diagnostics}(...)
}
\description{
Custom rbind implementation maintainig information about submodels
}
