% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{rdunif}
\alias{rdunif}
\title{Discrete uniform distribution allowing for varying lower and upper bounds.}
\usage{
rdunif(n, a, b)
}
\description{
Internal, should not be exported.
Based on https://stats.stackexchange.com/a/3939/73129
}
