% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{recompute_statistics-deprecated}
\alias{recompute_statistics-deprecated}
\title{Recompute SBC statistics without refitting models.}
\description{
Delegates directly to \code{recompute_SBC_statistics()}.
}
\seealso{
\code{\link{SBC-deprecated}}
}
\keyword{internal}
