% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{recompute_statistics}
\alias{recompute_statistics}
\title{Recompute SBC statistics without refitting models.}
\usage{
recompute_statistics(
  old_results,
  datasets,
  backend,
  thin_ranks = SBC_backend_default_thin_ranks(backend),
  gen_quants = NULL
)
}
\arguments{
\item{backend}{backend used to fit the results. Used to pull various defaults
and other setting influencing the computation of statistics.}
}
\value{
An S3 object of class \code{SBC_results} with updated \verb{$stats} and \verb{$default_diagnostics} fields.
}
\description{
Useful for example to recompute SBC ranks with a different choice of \code{thin_ranks}
or added generated quantities.
}
