% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_attributes.R
\name{remove_attribute_from_stats}
\alias{remove_attribute_from_stats}
\title{Remove an attribute from a \code{data.frame} column with attributes.}
\usage{
remove_attribute_from_stats(attribute, x)
}
\description{
Remove an attribute from a \code{data.frame} column with attributes.
}
\details{
By design, only one copy of an attribute is removed, if there are multiple copies
}
