% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backends.R
\name{rstan_sample_SBC_backend}
\alias{rstan_sample_SBC_backend}
\title{SBC backend using the \code{sampling} method from \code{rstan}.}
\usage{
rstan_sample_SBC_backend(model, ...)
}
\arguments{
\item{model}{a \code{stanmodel} object (created via \code{rstan::stan_model})}

\item{...}{other arguments passed to \code{sampling} (number of iterations, ...).
Arguments \code{data} and \code{cores} cannot be set this way as they need to be
controlled by the package.}
}
\description{
SBC backend using the \code{sampling} method from \code{rstan}.
}
