% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/self-calib.R
\name{sample_quantile_phi}
\alias{sample_quantile_phi}
\title{Given a vector of phis, which represent the quantiles of equally spaced probabilities on \link{0, 1} defined as i/S, return a function that returns N random samples from the quantile function}
\usage{
sample_quantile_phi(N, phis)
}
\arguments{
\item{N}{number of samples to draw, returned as a vector of length N}

\item{phis}{vector of phis to sample from}
}
\value{
a vector of samples drawn by inverse transform sampling
}
\description{
Given a vector of phis, which represent the quantiles of equally spaced probabilities on \link{0, 1} defined as i/S, return a function that returns N random samples from the quantile function
}
