% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend-bridgesampling.R
\name{SBC_backend_bridgesampling}
\alias{SBC_backend_bridgesampling}
\title{Backend using bridgesampling to do Bayes factor calculation (and Bayesian Model Averaging) over two candidate
models.}
\usage{
SBC_backend_bridgesampling(backend_H0, backend_H1, model_var = "model", ...)
}
\arguments{
\item{...}{passed to \code{\link[bridgesampling:bridge_sampler]{bridgesampling::bridge_sampler()}}.}
}
\description{
Backend using bridgesampling to do Bayes factor calculation (and Bayesian Model Averaging) over two candidate
models.
}
