% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backends.R
\name{SBC_backend_brms}
\alias{SBC_backend_brms}
\title{Build a backend based on the \code{brms} package.}
\usage{
SBC_backend_brms(..., template_dataset)
}
\arguments{
\item{...}{arguments passed to \code{brm}.}

\item{template_dataset}{a representative dataset that can be used to generate code.}
}
\description{
Build a backend based on the \code{brms} package.
}
