% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend-brms.R
\name{SBC_backend_brms_from_generator}
\alias{SBC_backend_brms_from_generator}
\title{Build a brms backend, reusing the compiled model from a previously created \code{SBC_generator_brms}
object.}
\usage{
SBC_backend_brms_from_generator(generator, ...)
}
\description{
Build a brms backend, reusing the compiled model from a previously created \code{SBC_generator_brms}
object.
}
