% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend-cached.R
\name{SBC_backend_cached}
\alias{SBC_backend_cached}
\title{A backend that wraps another backend with a caching scheme.}
\usage{
SBC_backend_cached(cache_dir, backend)
}
\arguments{
\item{cache_dir}{directory where the cache files (one file per fit) will
be stored}

\item{backend}{any other backend that does the actual computation}
}
\description{
The cache is stored in files named based on dataset and underlying
backend hashes, within the given directory, you can use
\code{\link[=cached_fit_filename]{cached_fit_filename()}} to manually access the cache.
}
\details{
The cached fit will try to re-emit all output, warnings and messages.
The ordering within each category is preserved, but the relative ordering of
e.g. outputs vs. messages (and all other combinations) may not be.
}
