% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend-cmdstanr.R
\name{SBC_backend_cmdstan_optimize}
\alias{SBC_backend_cmdstan_optimize}
\title{Backend based on optimizing via \code{cmdstanr}.}
\usage{
SBC_backend_cmdstan_optimize(model, ..., n_retries_init = 1)
}
\arguments{
\item{model}{an object of class \code{CmdStanModel} (as created by \code{cmdstanr::cmdstan_model})}

\item{...}{other arguments passed to the \verb{$optimize()} method of the model.
The \code{data} argument cannot be set this way as they need to be controlled by the SBC
package.}

\item{n_retries_init}{number of times to retry the optimizing fit if the algorithm
has trouble initializing (e.g. optimizing gets stuck)}
}
\description{
Backend based on optimizing via \code{cmdstanr}.
}
