% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backends.R
\name{SBC_backend_cmdstan_sample}
\alias{SBC_backend_cmdstan_sample}
\title{Backend based on sampling via \code{cmdstanr}.}
\usage{
SBC_backend_cmdstan_sample(model, ...)
}
\arguments{
\item{model}{an object of class \code{CmdStanModel} (as created by \code{cmdstanr::cmdstan_model})}

\item{...}{other arguments passed to the \verb{$sample()} method of the model. The \code{data} and
\code{parallel_chains} arguments cannot be set this way as they need to be controlled by the SBC
package.}
}
\description{
Backend based on sampling via \code{cmdstanr}.
}
