% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backends.R
\name{SBC_backend_default_thin_ranks}
\alias{SBC_backend_default_thin_ranks}
\alias{SBC_backend_default_thin_ranks.default}
\title{S3 generic to get backend-specific default thinning for rank computation.}
\usage{
SBC_backend_default_thin_ranks(backend)

\method{SBC_backend_default_thin_ranks}{default}(backend)
}
\description{
The default implementation plays it relatively safe and returns 10, unless
\code{\link[=SBC_backend_iid_samples]{SBC_backend_iid_samples()}} returns \code{TRUE} in which case it returns 1.
}
