% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend-BayesFactor.R
\name{SBC_backend_extractBF_comparison}
\alias{SBC_backend_extractBF_comparison}
\title{A backend for Bayes factor workflow with \link[BayesFactor:extractBF-methods]{BayesFactor::extractBF}.}
\usage{
SBC_backend_extractBF_comparison(backend_H0, backend_H1, model_var = "model")
}
\description{
A backend for Bayes factor workflow with \link[BayesFactor:extractBF-methods]{BayesFactor::extractBF}.
}
