% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backends.R
\name{SBC_backend_hash_for_cache}
\alias{SBC_backend_hash_for_cache}
\title{Get hash used to identify cached results.}
\usage{
SBC_backend_hash_for_cache(backend)
}
\description{
S3 generic that allows backends to override how a hash is computed. By default \code{rlang::hash()}
is used.
}
