% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backends.R
\name{SBC_backend_iid_samples}
\alias{SBC_backend_iid_samples}
\alias{SBC_backend_iid_samples.default}
\title{S3 generic to let backends signal that they produced independent samples.}
\usage{
SBC_backend_iid_samples(backend)

\method{SBC_backend_iid_samples}{default}(backend)
}
\arguments{
\item{backend}{to check}
}
\description{
Most backends (e.g. those based on variatns of MCMC) don't produce
independent samples and thus diagnostics like Rhat and ESS are important
and samples may need thinning. Backends that already produce independent
samples (e.g. ADVI/optimizing) can implement this method to return \code{TRUE}
to signal this is the case. If this method returns \code{TRUE}, ESS and Rhat will
always attain their best possible values and \code{\link[=SBC_backend_default_thin_ranks]{SBC_backend_default_thin_ranks()}}
will return \code{1}.
The default implementation returns \code{FALSE}.
}
