% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend-BayesFactor.R
\name{SBC_backend_lmBF}
\alias{SBC_backend_lmBF}
\title{A backend using \code{\link[BayesFactor:lmBF]{BayesFactor::lmBF()}}}
\usage{
SBC_backend_lmBF(..., iterations = 2000)
}
\arguments{
\item{...}{all parameters are passed directly to \code{\link[BayesFactor:lmBF]{BayesFactor::lmBF()}}.}
}
\description{
A backend using \code{\link[BayesFactor:lmBF]{BayesFactor::lmBF()}}
}
