% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend-mock.R
\name{SBC_backend_mock}
\alias{SBC_backend_mock}
\title{A mock backend.}
\usage{
SBC_backend_mock(
  result = posterior::draws_matrix(a = rnorm(100)),
  output = NULL,
  message = NULL,
  warning = NULL,
  error = NULL
)
}
\arguments{
\item{result}{a \code{draws_matrix} that will be returned regardless of the data}
}
\description{
Mock backend is useful for testing the package.
It will ignore all data passed to it and always
provide \code{result} as the draws generated by the backend.
}
