% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend-cached.R
\name{SBC_backend_postprocess_cached_fit}
\alias{SBC_backend_postprocess_cached_fit}
\title{Allows the backend to do any pre-/post- processing on a fit stored to / loaded from cache.}
\usage{
SBC_backend_postprocess_cached_fit(backend, generated, fit)
}
\description{
This is useful e.g. to provide a live \code{stanmodel} instance to an rstan fit,
as the one loaded from cache won't work for some cases.
}
\details{
The deafult implementation does nothing.
}
