% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend-rjags.R
\name{SBC_backend_rjags}
\alias{SBC_backend_rjags}
\title{Create a JAGS backend using \code{rjags}}
\usage{
SBC_backend_rjags(
  file,
  n.iter,
  n.burnin,
  variable.names,
  thin = 1,
  na.rm = TRUE,
  ...
)
}
\arguments{
\item{file}{model file or connection to model code (passed to \code{\link[rjags:jags.model]{rjags::jags.model()}})}

\item{n.iter}{number of iterations for sampling (passed to [rjags::coda.samples())}

\item{n.burnin}{number of iterations used for burnin}

\item{variable.names}{names of variables to monitor (passed to \code{\link[rjags:coda.samples]{rjags::coda.samples()}})}

\item{thin}{thinning (passed to \code{\link[rjags:coda.samples]{rjags::coda.samples()}})}

\item{na.rm}{whether to omit variables containing NA (passed to \code{\link[rjags:coda.samples]{rjags::coda.samples()}})}

\item{...}{additional arguments passed to \code{\link[rjags:jags.model]{rjags::jags.model()}}}
}
\description{
Create a JAGS backend using \code{rjags}
}
