% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backends.R
\name{SBC_backend_rstan_optimizing}
\alias{SBC_backend_rstan_optimizing}
\title{SBC backend using the \code{optimizing} method from \code{rstan}.}
\usage{
SBC_backend_rstan_optimizing(model, ...)
}
\arguments{
\item{model}{a \code{stanmodel} object (created via \code{rstan::stan_model})}

\item{...}{other arguments passed to \code{optimizing} (number of iterations, ...).
Argument \code{data} cannot be set this way as they need to be
controlled by the package.}
}
\description{
SBC backend using the \code{optimizing} method from \code{rstan}.
}
