% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backends.R
\name{SBC_backend_rstan_sample}
\alias{SBC_backend_rstan_sample}
\title{SBC backend using the \code{sampling} method from \code{rstan}.}
\usage{
SBC_backend_rstan_sample(model, ...)
}
\arguments{
\item{model}{a \code{stanmodel} object (created via \code{rstan::stan_model})}

\item{...}{other arguments passed to \code{sampling} (number of iterations, ...).
Arguments \code{data} and \code{cores} cannot be set this way as they need to be
controlled by the package.}
}
\description{
SBC backend using the \code{sampling} method from \code{rstan}.
}
