% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend-bridgesampling.R
\name{mutate.SBC_bridgesampling_diagnostics}
\alias{mutate.SBC_bridgesampling_diagnostics}
\alias{arrange.SBC_bridgesampling_diagnostics}
\alias{select.SBC_bridgesampling_diagnostics}
\title{Custom implementation of dplyr verbs to maintain  information about submodels}
\usage{
\method{mutate}{SBC_bridgesampling_diagnostics}(diags, ...)

\method{arrange}{SBC_bridgesampling_diagnostics}(diags, ...)

\method{select}{SBC_bridgesampling_diagnostics}(diags, ...)

\method{select}{SBC_bridgesampling_diagnostics}(diags, ...)
}
\description{
Custom implementation of dplyr verbs to maintain  information about submodels
}
