% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_factors.R
\name{SBC_datasets_for_bf}
\alias{SBC_datasets_for_bf}
\title{Combine two datasets to check Bayes factor computation.}
\usage{
SBC_datasets_for_bf(
  datasets_H0,
  datasets_H1,
  prob_H1 = 0.5,
  model_var = "model"
)
}
\description{
Merge two datasets of the same length, each generated by a different model. The "true" model
will be chosen randomly for each dataset.
}
\details{
The merged dataset will keep track of all parameters and later allow
splitting results with \code{\link[=split_SBC_results_for_bf]{split_SBC_results_for_bf()}}, to this, a number of
submodel-specific variables will be stored, but makred as
\code{\link[=hidden_var_attribute]{hidden_var_attribute()}} to not clutter default visualisations.
}
