% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{SBC_example_backend}
\alias{SBC_example_backend}
\title{Construct a backend to be used in the examples.}
\usage{
SBC_example_backend(
  example = c("normal_sd", "normal_var"),
  interface = c("rstan", "cmdstanr")
)
}
\arguments{
\item{example}{name of the example model}

\item{interface}{name of the interface to be used to fit the model}
}
\description{
Note that this will involve compiling a Stan model and may take a while.
}
