% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{SBC_example_generator}
\alias{SBC_example_generator}
\title{Construct a generator used in the examples.}
\usage{
SBC_example_generator(example = c("normal"), N = 100)
}
\arguments{
\item{example}{name of example}

\item{N}{size of the dataset the generator should simulate}
}
\value{
an object that can be passed to \code{\link[=generate_datasets]{generate_datasets()}}
}
\description{
Construct a generator used in the examples.
}
