% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{SBC_example_results}
\alias{SBC_example_results}
\title{Combine an example backend with an example generator to provide full
results that can be used to test other functions in the package.}
\usage{
SBC_example_results(
  example = c("normal_ok", "normal_bad", "visualizations"),
  interface = c("rstan", "cmdstanr", "rjags"),
  N = 100,
  n_sims = 50
)
}
\arguments{
\item{example}{\itemize{
\item name of the example. \code{normal_ok} is an example
where the generator matches the model
(using the \code{normal} generator and \code{normal_sd} backend), while
\code{normal_bad} is an example with a mismatch between the generator and backend
that manifests in SBC (\code{normal_bad} combines the \code{normal} generator with
\code{normal_bad} backend). \code{visualizations} creates a purely artificial results
that are meant to showcase the built-in plots (the \code{interface} parameter will
be ignored).
}}

\item{interface}{name of the interface to be used for the backend}

\item{N}{number of datapoints to simulate from the generator for each simulation}

\item{n_sims}{number of simulations to perform}
}
\description{
Except for \code{example = "visualizations"}, all examples will actually
compile and fit Stan models and thus may take a while to complete.
}
