% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backends.R
\name{SBC_fit_specific_dquants}
\alias{SBC_fit_specific_dquants}
\alias{SBC_fit_specific_dquants.default}
\title{A fit object may implement this method to add extra derived quantities
whose computation depend on some values only available in the fit.}
\usage{
SBC_fit_specific_dquants(fit)

\method{SBC_fit_specific_dquants}{default}(fit)
}
\value{
A \code{\link[=derived_quantities]{derived_quantities()}} object or NULL if there are none
}
\description{
As of 2026-03 the only use case is computing the top model probability
for model selection across multiple models (since model probability is not directly available
in normal derived quantities).
}
