% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend-BayesFactor.R
\name{SBC_fit_to_BFBayesFactor}
\alias{SBC_fit_to_BFBayesFactor}
\title{A fits need to implement this S3 generic to be useful for the
\code{\link[=SBC_backend_extractBF_comparison]{SBC_backend_extractBF_comparison()}} backend.}
\usage{
SBC_fit_to_BFBayesFactor(fit)
}
\description{
A fits need to implement this S3 generic to be useful for the
\code{\link[=SBC_backend_extractBF_comparison]{SBC_backend_extractBF_comparison()}} backend.
}
