% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backends.R
\name{SBC_fit_to_diagnostics}
\alias{SBC_fit_to_diagnostics}
\title{S3 generic to get backend-specific diagnostics.}
\usage{
SBC_fit_to_diagnostics(fit, fit_output, fit_messages)
}
\arguments{
\item{fit}{The fit returned by \code{SBC_fit}}

\item{fit_output}{a character string capturing what the backend wrote to stdout}

\item{fit_messages}{a \code{data.frame} with columns \code{type} (type of message as signalled by R,
currently either "message" or "warning") and \code{message} (the actual text of the message).}
}
\value{
an object that includes diagnostics or NULL, if no diagnostics available.
}
\description{
S3 generic to get backend-specific diagnostics.
}
