% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backends.R
\name{SBC_fit_to_draws_matrix}
\alias{SBC_fit_to_draws_matrix}
\alias{SBC_fit_to_draws_matrix.default}
\title{S3 generic converting a fitted model to a \code{draws_matrix} object.}
\usage{
SBC_fit_to_draws_matrix(fit)

\method{SBC_fit_to_draws_matrix}{default}(fit)
}
\description{
Needs to be implemented for all types of objects the backend can
return from \code{\link[=SBC_fit]{SBC_fit()}}. Default implementation just calls,
\code{\link[posterior:draws_matrix]{posterior::as_draws_matrix()}}, so if the fit object already supports
this, it will work out of the box.
}
