% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generator-brms.R
\name{SBC_generator_brms}
\alias{SBC_generator_brms}
\title{Create a brms generator.}
\usage{
SBC_generator_brms(
  formula,
  data,
  ...,
  generate_lp = TRUE,
  generate_lp_chunksize = 5000/nrow(data)
)
}
\arguments{
\item{formula, data, ...}{arguments passed to \code{brms::brm}}

\item{generate_lp}{whether to compute the overall log-likelihood of the model
as an additional variable. This can be somewhat computationally expensive,
but improves sensitivity of the SBC process.}

\item{generate_lp_chunksize}{Will determine whether
the chunk size used with \code{\link[future.apply:future_mapply]{future.apply::future_mapply()}}. Set quite high
by default as the parallelism only benefits for large individual datasets/number of
simulations.}
}
\description{
Brms generator uses a brms model with \code{sample_prior = "only"} to generate
new datasets.
}
