% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{SBC_generator_custom}
\alias{SBC_generator_custom}
\title{Wrap a function the creates a complete dataset.}
\usage{
SBC_generator_custom(f, ...)
}
\arguments{
\item{f}{function accepting at least an \code{n_datasets} argument and returning
and \code{SBC_datasets} object}

\item{...}{Additional arguments passed to \code{f}}
}
\description{
This creates a very thin wrapper around the function and can store additional
arguments, but does not do anything more..
}
\details{
Running:\if{html}{\out{<div class="r">}}\preformatted{gen <- SBC_generator_custom(f, <<some other args>>)
datasets <- generate_datasets(gen, n_datasets = my_n_datasets)
}\if{html}{\out{</div>}}

is equivalent to just running\if{html}{\out{<div class="r">}}\preformatted{datasets <- f(<<some other args>>, n_datasets = my_n_datasets)
}\if{html}{\out{</div>}}

So whenever you control the code calling \code{generate_datasets},
it usually makes more sense to just create an \code{SBC_datasets}
object directly and avoid using \code{SBC_generator_custom} and \code{generate_datasets} at all.
\code{SBC_generator_custom} can however be useful, when a code you
do not control calls \code{generate_datasets} for you and the
built-in generators do not provide you with enough flexibility.
}
