% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backends.R
\name{SBC_posterior_cdf}
\alias{SBC_posterior_cdf}
\alias{SBC_posterior_cdf.default}
\title{Use a fitted model to obtain a posterior CDF at a given point.}
\usage{
SBC_posterior_cdf(fit, variables)

\method{SBC_posterior_cdf}{default}(fit, variables)
}
\arguments{
\item{fit}{an object returned by the \code{\link[=SBC_fit]{SBC_fit()}} method.}

\item{variables}{a named vector of values at which to evaluate the CDF}
}
\value{
either \code{NULL} (the default implementation) or a \code{data.frame}
with a row for each variable. The columns must include:
\code{variable} (variable name), and either \code{cdf} (if there are no ties) or the
pair \code{cdf_low} and \code{cdf_high} indicate the possible tied values for the CDF.
}
\description{
Backends that represent explicit posterior distributions (i.e. not just samples)
can implement this S3 generic to evaluate the CDF of that distribution at
the simulated values.
}
\details{
A backend may choose to return explicit CDF only for some model parameters.
return from \code{\link[=SBC_fit]{SBC_fit()}}.
}
