% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{SBC_results}
\alias{SBC_results}
\title{SBC_results objects.}
\usage{
SBC_results(
  stats,
  fits,
  backend_diagnostics,
  default_diagnostics,
  outputs,
  messages,
  warnings,
  errors
)
}
\description{
The \code{SBC_results} contains the following fields:
\itemize{
\item \verb{$stats} statistics for all parameters and fits (one row per parameter-fit combination)
\item \verb{$fits}  the raw fits (unless \code{keep_fits = FALSE}) or \code{NULL} if the fit failed
\item \verb{$errors} error messages that caused fit failures
\item \verb{$outputs}, \verb{$messages}, \verb{$warnings} the outputs/messages/warnings written by fits
\item \verb{$default_diagnostics} a data frame of default convergence/correctness diagnostics (one row per fit)
\item \verb{$backend_diagnostics} a data frame of backend-specific diagnostics (one row per fit)
}
}
