% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{SBC_statistics_from_single_fit}
\alias{SBC_statistics_from_single_fit}
\title{Recompute SBC statistics given a single fit.}
\usage{
SBC_statistics_from_single_fit(
  fit,
  variables,
  generated,
  thin_ranks,
  ensure_num_ranks_divisor,
  dquants,
  backend,
  var_attributes = NULL,
  gen_quants = NULL
)
}
\arguments{
\item{thin_ranks}{how much thinning should be applied to posterior draws before computing
ranks for SBC. Should be large enough to avoid any noticeable autocorrelation of the
thinned draws See details below.}

\item{ensure_num_ranks_divisor}{Potentially drop some posterior samples to
ensure that this number divides the total number of SBC ranks (see Details).}

\item{dquants}{Derived quantities to include in SBC. Use \code{\link[=derived_quantities]{derived_quantities()}} to construct them.}

\item{backend}{the model + sampling algorithm. The built-in backends can be constructed
using \code{\link[=SBC_backend_cmdstan_sample]{SBC_backend_cmdstan_sample()}}, \code{\link[=SBC_backend_cmdstan_variational]{SBC_backend_cmdstan_variational()}},
\code{\link[=SBC_backend_rstan_sample]{SBC_backend_rstan_sample()}}, \code{\link[=SBC_backend_rstan_optimizing]{SBC_backend_rstan_optimizing()}} and \code{\link[=SBC_backend_brms]{SBC_backend_brms()}}.
(more to come: issue 31, 38, 39). The backend is an S3 class supporting at least the \code{\link[=SBC_fit]{SBC_fit()}},
\code{\link[=SBC_fit_to_draws_matrix]{SBC_fit_to_draws_matrix()}} methods.}

\item{gen_quants}{Deprecated, use dquants instead}
}
\description{
Potentially useful for doing some advanced stuff, but should not
be used in regular workflow. Use \code{\link[=recompute_SBC_statistics]{recompute_SBC_statistics()}} to update
an \verb{[SBC_results]} objects with different \code{thin_ranks} or other settings.
}
\seealso{
\code{\link[=recompute_SBC_statistics]{recompute_SBC_statistics()}}
}
