% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{SBCModel}
\alias{SBCModel}
\title{R6 Class Representing a collection of functions to extract data from a single model}
\description{
R6 Class Representing a collection of functions to extract data from a single model

R6 Class Representing a collection of functions to extract data from a single model
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{Some string to identify your SBCModel}

\item{\code{stan_model}}{A CmdStanModel or a stanmodel}

\item{\code{model_type}}{A string containing self$stan_model type. can be either RSTAN_MODEL_CLASS_NAME or CMDSTAN_MODEL_CLASS_NAME}

\item{\code{hyperpriors}}{A named list containing prior definitions and sampling functions
Initialize a SBCModel with a existing stan Model}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{SBCModel$new()}}
\item \href{#method-sample_theta_tilde}{\code{SBCModel$sample_theta_tilde()}}
\item \href{#method-sample_theta_tilde_stan}{\code{SBCModel$sample_theta_tilde_stan()}}
\item \href{#method-sample_y_tilde}{\code{SBCModel$sample_y_tilde()}}
\item \href{#method-sample_theta_bar_y}{\code{SBCModel$sample_theta_bar_y()}}
\item \href{#method-sample_bootstrap_y_tilde}{\code{SBCModel$sample_bootstrap_y_tilde()}}
\item \href{#method-clone}{\code{SBCModel$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SBCModel$new(name, stan_model = NULL, hyperpriors = list())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Some string to identify your SBCModel}

\item{\code{stan_model}}{A 'CmdStanModel' or a 'stanmodel' to draw samples}

\item{\code{hyperpriors}}{A named list containing definitions of prior for parameters. Follows the format:
list(param1=func, param2=func) where func is a function that returns a single value for each parameter}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sample_theta_tilde"></a>}}
\if{latex}{\out{\hypertarget{method-sample_theta_tilde}{}}}
\subsection{Method \code{sample_theta_tilde()}}{
Sample \eqn{\tilde{\theta} ~ P(\theta)} from self$hyperpriors
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SBCModel$sample_theta_tilde(pars, n_iters)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pars}}{List of parameters to draw prior samples.}

\item{\code{n_iters}}{Integer specifying number of draws.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Array of dimension(n_iters, n_pars) which are sampled prior values.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sample_theta_tilde_stan"></a>}}
\if{latex}{\out{\hypertarget{method-sample_theta_tilde_stan}{}}}
\subsection{Method \code{sample_theta_tilde_stan()}}{
Try and use the stan code to sample theta values
We'll set iteration step to 1 and sample from the hyperpriors.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SBCModel$sample_theta_tilde_stan(pars_list, n_iters)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pars_list}}{list of parameter names to draw}

\item{\code{n_iters}}{integer specifying number of individual draws

to be implemented(not sure if necessary)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sample_y_tilde"></a>}}
\if{latex}{\out{\hypertarget{method-sample_y_tilde}{}}}
\subsection{Method \code{sample_y_tilde()}}{
sample \eqn{\tilde{y} ~ p(\tilde{y} | \tilde{\theta})}, or y_tilde ~ P(y_tilde | theta_tilde)
The stan model must specify P(y | theta) within the generated quantities block
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SBCModel$sample_y_tilde(theta_arr, y_count, y_var = "y_rep", data = list())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{theta_arr}}{Array of sampled theta_tilde values (n_iters, n_pars), which is outputted from self$sample_theta_tilde()}

\item{\code{y_count}}{number of y_tilde samples being generated within generated quantities}

\item{\code{y_var}}{y_tilde variable name. Will be retrieved from model as so: y_var[n] (1<=n<=y_count)}

\item{\code{data}}{data list to pass to the stan model, in most cases would be dummy data since data will be drawn with fixed params}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
array of dimension (n_iters, y_count) of sampled y
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sample_theta_bar_y"></a>}}
\if{latex}{\out{\hypertarget{method-sample_theta_bar_y}{}}}
\subsection{Method \code{sample_theta_bar_y()}}{
sample \eqn{\theta ~ p(\theta | \tilde{y})}
The stan model must have \eqn{\theta} defined as parameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SBCModel$sample_theta_bar_y(
  y_sample_arr,
  data = list(),
  pars = list(),
  fit_iter = 200
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{y_sample_arr}}{array of dimension (n_iters, y_count) of sampled y. Same as output of self$sample_y_tilde()}

\item{\code{data}}{list of additional data to pass to the stan model. Note that "y" will be overwritten with draws from y_sample_arr.}

\item{\code{pars}}{list of parameters of interest.}

\item{\code{fit_iter}}{number of model iterations.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
: array of dimension (n_iters, n_pars, fit_iter) of posterior parameter draws
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sample_bootstrap_y_tilde"></a>}}
\if{latex}{\out{\hypertarget{method-sample_bootstrap_y_tilde}{}}}
\subsection{Method \code{sample_bootstrap_y_tilde()}}{
sample \eqn{\tilde{y}^{*} ~ Bootstrap(\tilde{y})}
given a single vector of data y_tilde, bootstrap sample multiple values of y_tilde_star[i] from bootstrap(y_tilde)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SBCModel$sample_bootstrap_y_tilde(y_sample_vector, n_iters = 2000)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{y_sample_vector}}{vector of y samples.}

\item{\code{n_iters}}{number of bootstrap samples to draw}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
array of dimension (n_iters, y_count) of bootstrap sampled y
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SBCModel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
