% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend-bridgesampling.R
\name{select.SBC_bridgesampling_diagnostics}
\alias{select.SBC_bridgesampling_diagnostics}
\title{Custom select implementation maintaining information about submodels}
\usage{
\method{select}{SBC_bridgesampling_diagnostics}(diags, ...)
}
\description{
Custom select implementation maintaining information about submodels
}
