% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/self-calib.R
\name{self_calib}
\alias{self_calib}
\title{Auto calibrate the initial prior samples using SBC iteration}
\usage{
self_calib(
  generator,
  backend,
  target_variables,
  bandwidth,
  mixture_mean_init,
  mixture_bw_init,
  thin,
  max_calib_iterations,
  sbc_iterations,
  fixed_generator_args
)
}
\arguments{
\item{generator}{function that generates datasets given each value in \code{param}}

\item{target_variables}{a named list defining parameter to mixture index mapping}

\item{bandwidth}{the smoothing bandwidth parameter to pass to stats.density}

\item{mixture_mean_init}{the initial mixture mean rvar}

\item{mixture_bw_init}{the initial mixture bandwidth rvar}

\item{thin}{Integer defining thinning parameter}

\item{max_calib_iterations}{the maximum number of iterations to run calibration. if not given will run indefinitely}

\item{sbc_iterations}{An integer returning the number of SBC iterations or a function with the signature (mixture_mean_rvar, mixture_bw_rvar) -> int
which returns the number of SBC iterations to run given previous calibration run results.}

\item{fixed_generator_args}{\emph{named list} containing additional arguments to generator}

\item{clamp_dist}{scalar clamp_disteter values e.g. prior_width}
}
\description{
Auto calibrate the initial prior samples using SBC iteration
}
