% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/self-calib.R
\name{self_calib_gmm}
\alias{self_calib_gmm}
\title{Auto calibrate the initial prior samples using SBC iteration and gmm approximation}
\usage{
self_calib_gmm(
  generator,
  backend,
  mixture_means_init_draws_rvars,
  mixture_bw_init_draws_rvars,
  nsims_fn,
  thin,
  max_selfcalib_iters,
  save_all_results,
  transform_types,
  fixed_generator_args
)
}
\arguments{
\item{generator}{function that generates datasets given each value in \code{param}}

\item{backend}{A backend object to use for running SBC}

\item{mixture_means_init_draws_rvars}{the initial mixture mean draws_rvars}

\item{mixture_bw_init_draws_rvars}{the initial mixture bandwidth draws_rvars}

\item{nsims_fn}{function with input: (mixture_means_rvar, mixture_bw_rvar), output: int
int is future number of parallel datasets to generate given true and its fitted hyperparameter (mixture_means)}

\item{thin}{Integer defining thinning parameter}

\item{max_selfcalib_iters}{the maximum number of iterations to run calibration. if not given will run indefinitely}

\item{save_all_results}{Boolean if TRUE returns a list of all SBC results, FALSE returns just the result of the last iteration.}

\item{transform_types}{Transformtype for mixture fitting}

\item{fixed_generator_args}{\emph{named list} containing additional arguments to pass to generator, \emph{after mixture_means_draws_rvars and mixture_bw_draws_rvars}}
}
\description{
Auto calibrate the initial prior samples using SBC iteration and gmm approximation
}
