% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric.R
\name{set2set}
\alias{set2set}
\title{Summarize relational property of overall prior and posterior samples}
\usage{
set2set(priors, posteriors, par, bins = 20)
}
\arguments{
\item{priors}{A posterior::draws_rvars of dimension(n_iterations=1, n_chains=n_sbc_iterations, n_variables=n_variables) which stores prior samples}

\item{posteriors}{A posterior::draws_Rvars of dimension(n_iterations=n_posterior_samples, n_chains=n_sbc_iterations, n_variables=n_variables), which stores fitted posterior samples}

\item{par}{names of parameter to summarize}

\item{bins}{number of bins for prior and post density}
}
\description{
Summarize relational property of overall prior and posterior samples
}
