% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_factors.R
\name{split_SBC_results_for_bf}
\alias{split_SBC_results_for_bf}
\title{Split stats from SBC of Bayes factor into stats for the individual submodels.}
\usage{
split_SBC_results_for_bf(results_bf)
}
\value{
list of two stats \code{data.frame}s, each for simulations where the
corresponding model was chosen.
}
\description{
Split stats from SBC of Bayes factor into stats for the individual submodels.
}
