% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{statistics_from_single_fit}
\alias{statistics_from_single_fit}
\title{Recompute SBC statistics given a single fit.}
\usage{
statistics_from_single_fit(fit, parameters, thin_ranks)
}
\description{
Potentially useful for doing some advanced stuff, but should not
be used in regular workflow. Use \code{recompute_statistics} to update
an \code{SBC_results} objects with different \code{thin_ranks} or other settings.
}
\seealso{
recompute_statistics
}
