% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{[.SBC_results}
\alias{[.SBC_results}
\title{Subset the results.}
\usage{
\method{[}{SBC_results}(x, indices)
}
\arguments{
\item{indices}{integer indices or a binary vector of the same length as the number fits,
selecting which fits to keep.}
}
\description{
Subset the results.
}
