% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{tf_param}
\alias{tf_param}
\title{Transform parameters from constrained to uncontrained}
\usage{
tf_param(param)
}
\arguments{
\item{param}{\code{draws_rvars} type parameter values}
}
\value{
list of uncontrained parameters and transformation type
}
\description{
Transform parameters from constrained to uncontrained
}
