% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/self-calib.R
\name{update_quantile_approximation}
\alias{update_quantile_approximation}
\title{Update mixture mean through quantile approxiation based on analytic gradient of quantile loss}
\usage{
update_quantile_approximation(
  hyperparam_rvar,
  hyperparam_hat_rvar,
  S,
  n_post_samples,
  epsilon
)
}
\arguments{
\item{hyperparam_rvar}{a posterior::rvar object of prior(pre-transformation) mixture mean values}

\item{hyperparam_hat_rvar}{a posterior::rvar object of posterior(post-transformation) mixture mean values}

\item{S}{Number of approximation points for the target quantile function.}

\item{n_post_samples}{the number of samples to draw from posterior, to approximate the expected quantile loss}

\item{epsilon}{gradient update coefficient}
}
\value{
a posterior::rvar object with the same dimension as the input rvars.
}
\description{
Update mixture mean through quantile approxiation based on analytic gradient of quantile loss
}
