% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backends.R
\name{validate_cdf_df}
\alias{validate_cdf_df}
\title{Validate the results of a SBC_posterior_cdf call.}
\usage{
validate_cdf_df(cdf_df, valid_variables)
}
\description{
Validate the results of a SBC_posterior_cdf call.
}
\keyword{internal}
