% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derived-quantities.R
\name{validate_derived_quantities}
\alias{validate_derived_quantities}
\title{Validate a definition of derived quantities evaluated in R.}
\usage{
validate_derived_quantities(x)
}
\description{
Validate a definition of derived quantities evaluated in R.
}
