% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric.R
\name{wasserstein}
\alias{wasserstein}
\title{wasserstein distance between binned samples}
\usage{
wasserstein(x, y)
}
\arguments{
\item{x}{numeric vector of density from first distribution}

\item{y}{numeric vector of density from second distribution}

\item{...}{unused}
}
\value{
distance value based on max difference
}
\description{
wasserstein distance between binned samples
}
