% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{attribute_present}
\alias{attribute_present}
\title{Check an attribute  in the \verb{$stats} element of the results
for presence of an attribute.}
\usage{
attribute_present(attribute, x)
}
\description{
Check an attribute  in the \verb{$stats} element of the results
for presence of an attribute.
}
