% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{calculate_ranks_draws_matrix}
\alias{calculate_ranks_draws_matrix}
\title{Calculate ranks given variable values within a posterior distribution.}
\usage{
calculate_ranks_draws_matrix(variables, dm, params = NULL)
}
\arguments{
\item{variables}{a vector of values to check}

\item{dm}{draws_matrix of the fit (assumed to be already thinned if that was necessary)}

\item{params}{DEPRECATED. Use \code{variables} instead.}
}
\description{
When there are ties (e.g. for discrete variables), the rank is currently drawn stochastically
among the ties. Depending on \code{na_lowest},
NA is assumed to be equal to any value.
}
