% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{check_all_SBC_diagnostics}
\alias{check_all_SBC_diagnostics}
\title{Check diagnostics and issue warnings when those fail.}
\usage{
check_all_SBC_diagnostics(x)
}
\description{
Check diagnostics and issue warnings when those fail.
}
