% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derived-quantities.R
\name{derived_quantities}
\alias{derived_quantities}
\title{Create a definition of derived quantities evaluated in R.}
\usage{
derived_quantities(..., .globals = list())
}
\arguments{
\item{...}{named expressions representing the quantitites}

\item{.globals}{A list of names of objects that are defined
in the global environment and need to present for the gen. quants. to evaluate.
It is added to the \code{globals} argument to \code{\link[future:future]{future::future()}}, to make those
objects available on all workers.}
}
\description{
When the expression contains non-library functions/objects, and parallel processing
is enabled, those must be
named in the \code{.globals} parameter (hopefully we'll be able to detect those
automatically in the future). Note that \code{\link[=recompute_SBC_statistics]{recompute_SBC_statistics()}} currently
does not use parallel processing, so \code{.globals} don't need to be set.
}
