% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_ecdf}
\alias{plot_ecdf}
\alias{plot_ecdf_diff}
\title{Plot the ECDF-based plots.}
\usage{
plot_ecdf(
  x,
  variables = NULL,
  K = NULL,
  gamma = NULL,
  prob = 0.95,
  size = 1,
  alpha = 0.33,
  ...,
  parameters = NULL
)

plot_ecdf_diff(
  x,
  variables = NULL,
  K = NULL,
  gamma = NULL,
  prob = 0.95,
  size = 1,
  alpha = 0.33,
  ...,
  parameters = NULL
)
}
\arguments{
\item{x}{object supporting the \code{\link[=data_for_ecdf_plots]{data_for_ecdf_plots()}} method.}

\item{variables}{optional subset of variables to show in the plot}

\item{K}{number of uniformly spaced evaluation points for the ECDF or ECDFs. Affects
the granularity of the plot and can significantly speed up the computation
of the simultaneous confidence bands. Defaults to the smaller of number of
ranks per variable and the maximum rank.}

\item{gamma}{TODO}

\item{prob}{the width of the plotted confidence interval for the ECDF.}

\item{size}{size passed to \code{\link[ggplot2:geom_ribbon]{ggplot2::geom_ribbon()}} for the confidence band}

\item{alpha}{alpha level of the confidence band}

\item{...}{additional arguments passed to \code{\link[=data_for_ecdf_plots]{data_for_ecdf_plots()}}.
Most notably, if \code{x} is matrix, a \code{max_rank} parameter needs to be given.}

\item{parameters}{DEPRECATED, use \code{variables} instead.}
}
\description{
See the methods for \code{\link[=data_for_ecdf_plots]{data_for_ecdf_plots()}} for available data formats.
}
\details{
\href{https://arxiv.org/abs/1903.08008}{arxiv::1903.08008} by A. Vehtari et al.
}
\seealso{
\code{\link[=plot_coverage]{plot_coverage()}}
}
