% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_ecdf}
\alias{plot_ecdf}
\alias{plot_ecdf_diff}
\title{Plot the ECDF-based plots.}
\usage{
plot_ecdf(
  x,
  parameters = NULL,
  K = NULL,
  gamma = NULL,
  prob = 0.95,
  size = 1,
  alpha = 0.33,
  ...
)

plot_ecdf_diff(
  x,
  parameters = NULL,
  K = NULL,
  gamma = NULL,
  prob = 0.95,
  size = 1,
  alpha = 0.33,
  ...
)
}
\arguments{
\item{x}{object supporting the \code{\link[=data_for_ecdf_plots]{data_for_ecdf_plots()}} method.}

\item{parameters}{optional subset of parameters to show in the plot}

\item{K}{number of uniformly spaced evaluation points for the ECDF or ECDFs. Affects
the granularity of the plot and can significantly speed up the computation
of the simultaneous confidence bands. Defaults to the smaller of number of
ranks per parameter and the maximum rank.}

\item{gamma}{TODO}

\item{prob}{the width of the plotted confidence interval for the ECDF.}

\item{size}{size passed to \code{\link[ggplot2:geom_ribbon]{ggplot2::geom_ribbon()}} for the confidence band}

\item{alpha}{alpha level of the confidence band}
}
\description{
\href{https://arxiv.org/abs/1903.08008}{arxiv::1903.08008} by A. Vehtari et al.
}
