% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary-calibration-tests.R
\name{gaffke_ci}
\alias{gaffke_ci}
\alias{gaffke_p}
\alias{gaffke_test}
\title{Non-parametric test for the mean of a bounded variable.}
\usage{
gaffke_ci(probs, B = 10000, alpha = 0.05)

gaffke_p(
  probs,
  mu = 0.5,
  alpha = 0.05,
  B = 10000,
  alternative = c("two.sided", "less", "greater")
)

gaffke_test(
  x,
  mu = 0.5,
  alpha = 0.05,
  lb = 0,
  ub = 1,
  B = 10000,
  alternative = c("two.sided", "less", "greater")
)
}
\arguments{
\item{B}{number of bootstrap samples for the null distribution}

\item{alpha}{the level of the test}

\item{mu}{the mean under null hypothesis}

\item{alternative}{the alternative for the test.}

\item{x}{a vector of observed values}

\item{lb}{the lower bound for \code{x}}

\item{ub}{the upper bound for \code{x}}
}
\value{
\code{gaffke_test} returns an object of class \code{htest}, \code{gaffke_p} and
\code{gaffke_ci} return just the p-value / CI as numeric for easier use in batch
workflows.
}
\description{
Test a null hypothesis about the mean of i.i.d. samples.
The test is based on Gaffke 2005, though a more detailed analysis and
exposition can be found in Learned-Miller & Thomas 2020.
}
\details{
The test is expected to be valid for any bounded distribution without further
assumptions. The test has been proven valid only for special cases but
no counterexample is known despite some efforts in the literature to find
some.
}
\references{
Gaffke, N. (2005).
“Three test statistics for a nonparametric one-sided hypothesis on the mean
of a nonnegative variable.” Mathematical Methods of Statistics, 14(4): 451–467.

Learned-Miller, E. and Thomas, P. S. (2020).
“A New Confidence Interval for the Mean of a Bounded Random Variable.”
https://arxiv.org/abs/1905.06208
}
