% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{generate_datasets}
\alias{generate_datasets}
\title{Generate datasets.}
\usage{
generate_datasets(generator, n_datasets)
}
\arguments{
\item{generator}{a generator object - build e.g. via \code{SBC_generator_function} or
\code{SBC_generator_brms}.}
}
\value{
object of class \code{SBC_datasets}
TODO: seed
}
\description{
Generate datasets.
}
