% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_contraction}
\alias{plot_contraction}
\title{Prior/posterior contraction plot.}
\usage{
plot_contraction(
  x,
  prior_sd,
  variables = NULL,
  scale = "sd",
  alpha = 0.8,
  parameters = NULL
)
}
\arguments{
\item{x}{object containing results (a data.frame or \code{\link[=SBC_results]{SBC_results()}} object).}

\item{prior_sd}{a named vector of prior standard deviations for your variables.
Either pass in analytically obtained values or use \code{\link[=calculate_prior_sd]{calculate_prior_sd()}} to get an empirical estimate from
an \code{SBC_datasets} object.}

\item{variables}{variables to show in the plot or \code{NULL} to show all
must correspond a field already computed in the results (most likely \code{"mean"} and \code{"median"}).}

\item{scale}{which scale of variability you want to see - either \code{"sd"} for standard deviation
or \code{"var"} for variance.}

\item{alpha}{the alpha for the points}

\item{parameters}{DEPRECATED, use \code{variables} instead.}
}
\value{
a ggplot2 plot object
}
\description{
The rationale for this plot and its interpretation is explained in
Mike Betancourt's
\href{https://betanalpha.github.io/assets/case_studies/principled_bayesian_workflow.html#132_A_Bayesian_Eye_Chart}{Towards A Principled Bayesian Workflow}.
}
