% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_coverage}
\alias{plot_coverage}
\alias{plot_coverage.SBC_results}
\alias{plot_coverage.data.frame}
\title{Plot the observed coverage and its uncertainty}
\usage{
plot_coverage(x, parameters = NULL, prob = 0.95, interval_type = "central")

\method{plot_coverage}{SBC_results}(x, parameters = NULL, prob = 0.95, interval_type = "central")

\method{plot_coverage}{data.frame}(x, parameters = NULL, prob = 0.95, interval_type = "central")
}
\arguments{
\item{x}{object containing results (a data.frame or \code{\link[=SBC_results]{SBC_results()}} object).}

\item{parameters}{parameters to show in the plot or \code{NULL} to show all}

\item{prob}{the with of the uncertainty interval to be shown}
}
\value{
a ggplot2 plot object
}
\description{
Please refer to \code{\link[=observed_coverage]{observed_coverage()}} for details on computation
and limitations of this plot as well as details on the arguments.
}
