% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_rank_hist}
\alias{plot_rank_hist}
\title{Plot rank histogram of an SBC results.}
\usage{
plot_rank_hist(
  x,
  variables = NULL,
  bins = NULL,
  prob = 0.95,
  ...,
  parameters = NULL
)
}
\arguments{
\item{x}{Object supporting the plotting method.}
}
\description{
By default the support is for \code{SBC_results} objects and data frames in the same
format as the \verb{$stats} element of \code{SBC_results}.
}
