% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{recompute_SBC_statistics}
\alias{recompute_SBC_statistics}
\title{Recompute SBC statistics without refitting models.}
\usage{
recompute_SBC_statistics(
  old_results,
  datasets,
  backend,
  thin_ranks = SBC_backend_default_thin_ranks(backend),
  ensure_num_ranks_divisor = 2,
  dquants = NULL,
  gen_quants = NULL
)
}
\arguments{
\item{datasets}{an object of class \code{SBC_datasets}}

\item{backend}{backend used to fit the results. Used to pull various defaults
and other setting influencing the computation of statistics.}

\item{thin_ranks}{how much thinning should be applied to posterior draws before computing
ranks for SBC. Should be large enough to avoid any noticeable autocorrelation of the
thinned draws See details below.}

\item{ensure_num_ranks_divisor}{Potentially drop some posterior samples to
ensure that this number divides the total number of SBC ranks (see Details).}

\item{dquants}{Derived quantities to include in SBC. Use \code{\link[=derived_quantities]{derived_quantities()}} to construct them.}

\item{gen_quants}{Deprecated, use dquants instead}
}
\value{
An S3 object of class \code{SBC_results} with updated \verb{$stats} and \verb{$default_diagnostics} fields.
}
\description{
Useful for example to recompute SBC ranks with a different choice of \code{thin_ranks}
or added derived quantities.
}
