% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{recompute_statistics}
\alias{recompute_statistics}
\title{Recompute SBC statistics without refitting models.}
\usage{
recompute_statistics(old_results, datasets, thin_ranks = 10, gen_quants = NULL)
}
\value{
An S3 object of class \code{SBC_results} with updated \verb{$stats} and \verb{$default_diagnostics} fields.
}
\description{
Useful for example to recompute SBC ranks with a different choice of \code{thin_ranks}
or added generated quantities.
}
