% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend-brms.R
\name{SBC_backend_brms}
\alias{SBC_backend_brms}
\title{Build a backend based on the \code{brms} package.}
\usage{
SBC_backend_brms(
  ...,
  template_data,
  out_stan_file = NULL,
  template_dataset = NULL
)
}
\arguments{
\item{...}{arguments passed to \code{brm}.}

\item{template_data}{a representative value for the \code{data} argument in \code{brm}
that can be used to generate code.}

\item{out_stan_file}{A filename for the generated Stan code. Useful for
debugging and for avoiding unnecessary recompilations.}

\item{template_dataset}{DEPRECATED. Use \code{template_data}}
}
\description{
Build a backend based on the \code{brms} package.
}
