% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend-cmdstanr.R
\name{SBC_backend_cmdstan_sample}
\alias{SBC_backend_cmdstan_sample}
\title{Backend based on sampling via \code{cmdstanr}.}
\usage{
SBC_backend_cmdstan_sample(model, ..., init_factory = NULL)
}
\arguments{
\item{model}{an object of class \code{CmdStanModel} (as created by \code{cmdstanr::cmdstan_model})}

\item{...}{other arguments passed to the \verb{$sample()} method of the model. The \code{data} and
\code{parallel_chains} arguments cannot be set this way as they need to be controlled by the SBC
package.}

\item{init_factory}{an optional function that takes in a dataset and returns a value that
can be passed to the \code{init} argument of \verb{$sample()}. This allows for data-dependent inits.
The caching mechanism in \code{\link[=compute_SBC]{compute_SBC()}} ignores the environment of the function, i.e.
if the init factory takes values from its environment and those values change between
runs, this will not by itself cause cached results to be recomputed.}
}
\description{
Backend based on sampling via \code{cmdstanr}.
}
