% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backends.R
\name{SBC_backend_cmdstan_variational}
\alias{SBC_backend_cmdstan_variational}
\title{Backend based on variational approximation via \code{cmdstanr}.}
\usage{
SBC_backend_cmdstan_variational(model, ...)
}
\arguments{
\item{model}{an object of class \code{CmdStanModel} (as created by \code{cmdstanr::cmdstan_model})}

\item{...}{other arguments passed to the \verb{$variational()} method of the model.
The \code{data} argument cannot be set this way as they need to be controlled by the SBC
package.}
}
\description{
Backend based on variational approximation via \code{cmdstanr}.
}
