% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend-rstan.R
\name{SBC_backend_rstan_optimizing}
\alias{SBC_backend_rstan_optimizing}
\title{SBC backend using the \code{optimizing} method from \code{rstan}.}
\usage{
SBC_backend_rstan_optimizing(model, ..., n_retries_hessian = 1)
}
\arguments{
\item{model}{a \code{stanmodel} object (created via \code{rstan::stan_model})}

\item{...}{other arguments passed to \code{optimizing} (number of iterations, ...).
Argument \code{data} cannot be set this way as they need to be
controlled by the package.}

\item{n_retries_hessian}{the number of times the backend is allow to retry optimization
(with different seeed) to produce a usable Hessian that can produce draws. In some cases,
the Hessian may be numerically unstable and not be positive definite.}
}
\description{
SBC backend using the \code{optimizing} method from \code{rstan}.
}
