% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend-bridgesampling.R
\name{SBC_fit_to_bridge_sampler}
\alias{SBC_fit_to_bridge_sampler}
\title{Convert a fit for a given backend into a bridge sampler object.}
\usage{
SBC_fit_to_bridge_sampler(backend, fit, generated, ...)
}
\arguments{
\item{...}{passed to \code{\link[bridgesampling:bridge_sampler]{bridgesampling::bridge_sampler()}}.}
}
\description{
Convert a fit for a given backend into a bridge sampler object.
}
\seealso{
\code{\link[bridgesampling:bridge_sampler]{bridgesampling::bridge_sampler()}}
}
