% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backends.R
\name{SBC_fit_to_diagnostics}
\alias{SBC_fit_to_diagnostics}
\title{S3 generic to get backend-specific diagnostics.}
\usage{
SBC_fit_to_diagnostics(fit, fit_output, fit_messages, fit_warnings)
}
\arguments{
\item{fit}{The fit returned by \code{SBC_fit}}

\item{fit_output}{a character string capturing what the backend wrote to stdout}

\item{fit_messages}{a character vector of messages the backend raised}

\item{fit_warnings}{a character vector of warnings the backend raised}
}
\value{
an single row \code{data.frame} that includes diagnostics or NULL, if no diagnostics available.
}
\description{
The diagnostics object has to be a \code{data.frame} but may
inherit additional classes - in particular it may be useful
for the returning object to implement \code{\link[=get_diagnostic_messages]{get_diagnostic_messages()}}.
}
