% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{SBC_print_example_model}
\alias{SBC_print_example_model}
\title{Print the Stan code of a model used in the examples.}
\usage{
SBC_print_example_model(example = c("normal_sd", "normal_var"))
}
\arguments{
\item{example}{name of the example model.}
}
\description{
Print the Stan code of a model used in the examples.
}
