% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{SBC_results}
\alias{SBC_results}
\title{Create an \code{SBC_results} object}
\usage{
SBC_results(
  stats,
  fits,
  backend_diagnostics,
  default_diagnostics,
  outputs,
  messages,
  warnings,
  errors
)
}
\description{
This will build and validate an \code{SBC_results} object from its constituents.
}
\details{
The \code{SBC_results} contains the following fields:
\itemize{
\item \verb{$stats} statistics for all variables and fits (one row per variable-fit combination)
\item \verb{$fits}  the raw fits (unless \code{keep_fits = FALSE}) or \code{NULL} if the fit failed
\item \verb{$errors} error messages that caused fit failures
\item \verb{$outputs}, \verb{$messages}, \verb{$warnings} the outputs/messages/warnings written by fits
\item \verb{$default_diagnostics} a data frame of default convergence/correctness diagnostics (one row per fit)
\item \verb{$backend_diagnostics} a data frame of backend-specific diagnostics (one row per fit)
}
}
