% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/self-calib.R
\name{self_calib}
\alias{self_calib}
\title{Auto calibrate the initial prior samples using SBC iteration}
\usage{
self_calib(
  generator,
  hyperparam,
  param,
  predictor,
  backend,
  target_vars,
  dist,
  dist_init = 0,
  thin,
  cnt = 0,
  evolve_df,
  delivDir
)
}
\arguments{
\item{generator}{function that generates datasets given each value in \code{param}}

\item{hyperparam}{scalar hyperparameter values e.g. prior_width}

\item{param}{rvars<S>\link{N} prior values i.e. parameter values to be tested "draws_rvars"}

\item{predictor}{rvars<N>\link{S} predictor values}

\item{backend}{stan_model that samples posterior with \code{datasets} simulated from the former two}

\item{target_vars}{function of parameters with which SBC iteration convergence are judged}

\item{dist}{empirical distance between the previous and the current parameter values, default is \code{cjs_dist}}

\item{thin}{scalar select 1 out of \code{thin} number of samples, \code{iter_sampling}=\code{thin}*M, default: 10}

\item{cnt}{function of parameters with which SBC iteration convergence are judged}

\item{evolve_df}{datafraame holding \verb{median, mad (or sd)} of samples from every iteration}

\item{delivDir}{save location of output result}
}
\value{
next param values summarized from \code{S} * \code{M} posterior samples
}
\description{
Auto calibrate the initial prior samples using SBC iteration
}
