% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/self-calib.R
\name{self_calib_adaptive}
\alias{self_calib_adaptive}
\title{Auto calibrate the initial prior samples using SBC iterations, with an adaptive update strategy}
\usage{
self_calib_adaptive(
  generator,
  backend,
  updator,
  target_params,
  init_lambdas,
  nsims,
  gamma,
  tol,
  fixed_args
)
}
\arguments{
\item{generator}{function that generates datasets given each value in \code{param}}

\item{backend}{backend object to use for running SBC}

\item{updator}{hyperparameter update type}

\item{nsims}{number of datasets i.e. prior draws}

\item{gamma}{convergence speed e.g. step size}

\item{tol}{tolerence for determining termination}

\item{fixed_args}{\emph{named list} containing additional arguments to pass to generator, \emph{after mu and sigma}}

\item{target_param}{list of strings indicating target parameter names}

\item{init_mu}{initial lambda_mu value to use}

\item{init_sigma}{initial lambda_sigma value to use}
}
\description{
Auto calibrate the initial prior samples using SBC iterations, with an adaptive update strategy
}
