% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/self-calib.R
\name{self_calib_adaptive}
\alias{self_calib_adaptive}
\title{Auto calibrate the initial prior samples using SBC iterations, with an adaptive update strategy}
\usage{
self_calib_adaptive(
  generator,
  approximator,
  sens_stan_model_dir,
  target_param,
  init_mu,
  init_sigma,
  nsims,
  ndraws,
  nchains,
  tol,
  fixed_args
)
}
\arguments{
\item{generator}{function that generates datasets given each value in \code{param}}

\item{approximator}{A string, either 'sampling' or 'optimizing', which defines the inference method}

\item{target_param}{target parameter name}

\item{init_mu}{initial lambda_mu value to use}

\item{init_sigma}{initial lambda_sigma value to use}

\item{nsims}{number of simulations?}

\item{ndraws}{number of posterior draws}

\item{nchains}{number of chains}

\item{tol}{tolerence for determining termination}

\item{fixed_args}{\emph{named list} containing additional arguments to pass to generator, \emph{after mu and sigma}}
}
\description{
Auto calibrate the initial prior samples using SBC iterations, with an adaptive update strategy
}
