% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{statistics_from_single_fit}
\alias{statistics_from_single_fit}
\title{Recompute SBC statistics given a single fit.}
\usage{
statistics_from_single_fit(
  fit,
  variables,
  generated,
  thin_ranks,
  gen_quants,
  backend
)
}
\description{
Potentially useful for doing some advanced stuff, but should not
be used in regular workflow. Use \code{\link[=recompute_SBC_statistics]{recompute_SBC_statistics()}} to update
an \verb{[SBC_results]} objects with different \code{thin_ranks} or other settings.
}
\seealso{
\code{\link[=recompute_SBC_statistics]{recompute_SBC_statistics()}}
}
