% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{tf_param_vec}
\alias{tf_param_vec}
\title{Transform parameters from constrained to uncontrained}
\usage{
tf_param_vec(param)
}
\arguments{
\item{param}{a vector}
}
\value{
list containing uncontrained parameters and transformation type
}
\description{
Transform parameters from constrained to uncontrained
}
