% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_attributes.R
\name{var_attributes}
\alias{var_attributes}
\alias{validate_var_attributes}
\alias{possibly_constant_var_attribute}
\alias{binary_var_attribute}
\alias{hidden_var_attribute}
\alias{na_valid_var_attribute}
\alias{inf_valid_var_attribute}
\alias{submodel_var_attribute}
\title{Predefined constants to use for variable attributes recognized by SBC.}
\usage{
var_attributes(...)

validate_var_attributes(var_attr)

possibly_constant_var_attribute()

binary_var_attribute()

hidden_var_attribute()

na_valid_var_attribute()

inf_valid_var_attribute()

submodel_var_attribute(sub_id)
}
\description{
Attributes give additional information useful for presenting SBC results
concerning the variables.
}
\details{
Should be passed via an instance of the \code{var_attributes} class to
\code{\link[=SBC_datasets]{SBC_datasets()}} (e.g. by returning a \verb{$var_attributes} element from a
function passed to \code{\link[=SBC_generator_function]{SBC_generator_function()}}).

\code{possibly_constant_var_attribute} attribute signals that having all
posterior draws identical is possible and thus no warnings should be
made for the resulting NAs in rhat and ESS checks.

\code{binary_var_attribute} marks the attribute as a binary variable (0 or 1)
and thus eligible for some special handling and visualisations
(e.g., \code{\link[=binary_probabilites_from_stats]{binary_probabilites_from_stats()}}, \code{\link[=binary_calibration_from_bp]{binary_calibration_from_bp()}}).

\code{hidden_var_attribute} will hide the variable in default visualisations,
unless the variable is explicitly mentioned.

\code{na_valid_var_attribute} will treat NAs as potentially equal to any
other value in rank ordering. This gives the expected results when NA
represents rare problems in computation that
should be ignored (see \code{\link[=calculate_ranks_draws_matrix]{calculate_ranks_draws_matrix()}}). Setting this
attribute also changes the ESS/Rhat computation to ignore NAs.

\code{inf_valid_var_attribute} means infinity values may appear in the samples
(this is useful e.g. to note that the parameter is actually not present for the
given draw). Setting this
attribute also changes the ESS/Rhat computation to ignore infinities.

\code{submodel_var_attribute} signals that the parameter belongs to a submodel
which can be extracted individually

In SBC results, the attributes of a variable are summarised in the
\code{attributes} column of the \verb{$stats} data.frame. Use \code{\link[=attribute_present_stats]{attribute_present_stats()}}
to check for presence of an attribute there.
}
