% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{possibly_constant_var_attribute}
\alias{possibly_constant_var_attribute}
\alias{binary_var_attribute}
\alias{hidden_var_attribute}
\title{Predefined constants to use for variable attributes recognized by SBC.}
\usage{
possibly_constant_var_attribute()

binary_var_attribute()

hidden_var_attribute()
}
\description{
Attributes give additional information useful for presenting SBC results
concerning the variables.
}
\details{
Should be passed via a named list to the \code{var_attributes} argument of
\code{\link[=SBC_datasets]{SBC_datasets()}} (e.g. by returning a \verb{$var_attributes} element from a
function passed to \code{\link[=SBC_generator_function]{SBC_generator_function()}}).

\code{possibly_constant_var_attribute} attribute signals that having all
posterior draws identical is possible and thus no warnings should be
made for the resulting NAs in rhat and ESS checks.

\code{binary_var_attribute} marks the attribute as a binary variable (0 or 1)
and thus eligible for some special visualisations (TODO).

\code{hidden_var_attribute} will hide the variable in default visualisations,
unless the variable is explicitly mentioned.

In SBC results, the attributes of a variable are summarised in the
\code{attributes} column of the \verb{$stats} data.frame. Use \code{\link[=attribute_present]{attribute_present()}}
to check for presence of an attribute there.
}
